Separar el sudoku en otro componente. crear botones de dificultad y de volver atras
