+++ README.md
@@ -2,4 +2,7 @@
 Esta es mi versión del famoso juego de sudokus
 
 
+Realizado con React
+Utiliza librería de React-Bootstrap
+
 ### `npm run start`
