+++ src/App.css
@@ -1,46 +1,22 @@
-
-.sudoku-board {
-  margin: 10px;
-  display: grid;
-  grid-template-columns: repeat(9, 40px);
-  grid-template-rows: repeat(9, 40px); 
-  gap: 1px;
-  width: 368px;
-  height: 366px;
-  background: black;
-  padding: 3px;
-  
+h1 {
+  text-align: center;
+  margin-top: 50px;
+  color: rgb(31, 11, 130);
+  width: 30%;
+  margin-inline: auto;
+  font-family: 'Cantarell', sans-serif;
 }
 
-.sudoku-cell {
+
+
+
+.buttonGroup {
   display: flex;
   justify-content: center;
   align-items: center;
-  font-size: 20px;
-  background: white;
-  border: none;
-  text-align: center;
-  color:blue;
-}
-.sudoku-cell:hover {
-  background: rgb(214, 233, 240);
-} */
-.sudoku-cell:focus {
-  outline: none;
-  background: rgb(214, 233, 240);
-}
-.sudoku-cell:disabled {
-  color: black;
-  background: white;
+  gap: 10px;
 }
 
-
-/* Estilos especiales para los bloques de 3x3*/
- .sudoku-cell:nth-child(3n):not(:nth-child(9n)) {
-  border-right: 2px solid black;
-}
-.sudoku-cell:nth-child(27n), .sudoku-cell:nth-child(27n-1), .sudoku-cell:nth-child(27n-2),
-.sudoku-cell:nth-child(27n-3), .sudoku-cell:nth-child(27n-4), .sudoku-cell:nth-child(27n-5),
-.sudoku-cell:nth-child(27n-6), .sudoku-cell:nth-child(27n-7), .sudoku-cell:nth-child(27n-8){
-  border-bottom: 2px solid black;
+.difficulty-buttom {
+  background-color: rgb(226, 162, 43);
 }
