+++ src/App.js
@@ -7,6 +7,7 @@ function App() {
 
   return (
     <div className="App">
+      <h1>SUDOKU</h1>
       <Home/>
     </div>
   );
