+++ src/index.css
@@ -1,5 +1,6 @@
 body {
   margin: 0;
+  background-color: rgba(155, 244, 201, 0.849);
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
